+++ src/electron-locator.ts
